# resources/docs/blog/openapi/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Swagger/OpenAPI documentation for the backend service.",
        "title": "Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/auth/password/reset": {
            "post": {
                "description": "set a new password using reset token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Reset password",
                "parameters": [
                    {
                        "description": "Reset password data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/resetpassword.Request"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/auth/token/refresh": {
            "post": {
                "description": "exchange refresh token for new access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh token",
                "parameters": [
                    {
                        "description": "Refresh request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/refresh.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/refresh.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/comments": {
            "get": {
                "description": "list comments optionally filtered by object and paginated",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "List comments",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Object UUID",
                        "name": "object_uuid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Object type",
                        "name": "object_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_khanzadimahdi_testproject_application_comment_getComments.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "create a comment on an object (article, etc.) as the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Create a new comment",
                "parameters": [
                    {
                        "description": "Comment payload",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_khanzadimahdi_testproject_application_comment_createComment.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/dashboard/articles": {
            "get": {
                "description": "page through articles in dashboard",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard articles"
                ],
                "summary": "List dashboard articles",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_khanzadimahdi_testproject_application_dashboard_article_getArticles.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "update existing article fields, current user must be author",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard articles"
                ],
                "summary": "Update a dashboard article",
                "parameters": [
                    {
                        "description": "Article update payload",
                        "name": "article",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updatearticle.Request"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "create a new article in dashboard as current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard articles"
                ],
                "summary": "Create dashboard article",
                "parameters": [
                    {
                        "description": "Article data",
                        "name": "article",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/createarticle.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/createarticle.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/dashboard/articles/{uuid}": {
            "get": {
                "description": "retrieve one article by UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard articles"
                ],
                "summary": "Get dashboard article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_khanzadimahdi_testproject_application_dashboard_article_getArticle.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "remove an article by UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard articles"
                ],
                "summary": "Delete dashboard article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "createarticle.Request": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "excerpt": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "video": {
                    "type": "string"
                }
            }
        },
        "createarticle.Response": {
            "type": "object",
            "properties": {
                "errors": {
                    "$ref": "#/definitions/domain.ValidationErrors"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "domain.ValidationErrors": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
...
        "getarticle.author": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "getarticles.author": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "getarticles.pagination": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_khanzadimahdi_testproject_application_comment_createComment.Request": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "object_type": {
                    "type": "string"
                },
                "object_uuid": {
                    "type": "string"
                },
                "parent_uuid": {
                    "type": "string"
                }
            }
        },
        "github_com_khanzadimahdi_testproject_application_comment_getComments.Response": {
            "type": "object",
            "properties": {
                "errors": {
                    "$ref": "#/definitions/domain.ValidationErrors"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_khanzadimahdi_testproject_application_comment_getComments.commentResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_khanzadimahdi_testproject_application_comment_getComments.pagination"
                }
            }
        },
        "github_com_khanzadimahdi_testproject_application_comment_getComments.authorResponse": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "github_com_khanzadimahdi_testproject_application_comment_getComments.commentResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/github_com_khanzadimahdi_testproject_application_comment_getComments.authorResponse"
                },
                "body": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "parent_uuid": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "github_com_khanzadimahdi_testproject_application_comment_getComments.pagination": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_khanzadimahdi_testproject_application_dashboard_article_getArticle.Response": {
            "type": "object",
            "properties": {
                "avatar": {
                    "$ref": "#/definitions/getarticle.author"
                },
                "body": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "excerpt": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "video": {
                    "type": "string"
                },
                "view_count": {
                    "type": "integer"
                }
            }
        },
        "github_com_khanzadimahdi_testproject_application_dashboard_article_getArticles.Response": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_khanzadimahdi_testproject_application_dashboard_article_getArticles.articleResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/getarticles.pagination"
                }
            }
        },
        "github_com_khanzadimahdi_testproject_application_dashboard_article_getArticles.articleResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/getarticles.author"
                },
                "cover": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "video": {
                    "type": "string"
                }
            }
        },
...
        "refresh.Request": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "refresh.Response": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "errors": {
                    "$ref": "#/definitions/domain.ValidationErrors"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
        "resetpassword.Request": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "updatearticle.Request": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "excerpt": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "video": {
                    "type": "string"
                }
            }
        },
...
    }
}